NeuroExpert-main/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
sentry-sdk>=1.40.0
pytest>=7.4.4
httpx>=0.26.0
pytest-cov>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
asyncpg>=0.29.0
greenlet>=3.0.3
